requirements.txt
scipy
Pillow
imutils
numpy
pygame
opencv-python
opencv-contrib-python
cmake
dlib
